type mismatch

+++ cs336_systems/flash_attention/flash_attention_triton.py
@@ -232,7 +232,7 @@ def flash_bwd_kernel(
     dOi = tl.load(dO_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
     Li = tl.load(L_block_ptr, boundary_check=(0,), padding_option="zero").to(tl.float16)
     
-    dQi = tl.zeros((Q_TILE_SIZE, D), dtype=tl.float32)
+    dQi = tl.zeros((Q_TILE_SIZE, D), dtype=tl.float16)
 
     for j in range(tl.cdiv(N_KEYS, K_TILE_SIZE)):
         Kj = tl.load(K_block_ptr, boundary_check=(0, 1), padding_option="zero").to(tl.float16)
